requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
duplicity==2.1.4
exceptiongroup==1.2.0
fasteners==0.18
freetype-py==2.4.0
frozenlist==1.5.0
h11==0.14.0
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.3.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
magic-filter==1.0.12
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
multidict==6.1.0
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
outcome==1.2.0
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
propcache==0.2.1
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rlPyCairo==0.3.0
SecretStorage==3.3.3
selenium==4.18.1
setuptools==68.1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==235
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
webdriver-manager==4.0.2
websockets==10.4
wheel==0.42.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
yt-dlp==2024.4.9
zipp==1.0.0
